refactored teamSlackChannelEvent changes
