+++ nucleus/src/main/java/za/co/absa/subatomic/application/team/TeamService.java
@@ -7,20 +7,18 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
-import lombok.Value;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.client.RestTemplate;
 import za.co.absa.subatomic.adapter.team.rest.MembershipRequestResource;
 import za.co.absa.subatomic.domain.exception.DuplicateRequestException;
 import za.co.absa.subatomic.domain.exception.InvalidRequestException;
+import za.co.absa.subatomic.domain.member.TeamMemberSlackIdentity;
 import za.co.absa.subatomic.domain.team.MembershipRequestStatus;
-import za.co.absa.subatomic.infrastructure.AtomistConfigurationProperties;
-import za.co.absa.subatomic.infrastructure.atomist.resource.AtomistMemberBase;
+import za.co.absa.subatomic.domain.team.TeamSlackIdentity;
 import za.co.absa.subatomic.infrastructure.atomist.resource.team.AtomistTeam;
+import za.co.absa.subatomic.infrastructure.atomist.resource.AtomistMemberBase;
 import za.co.absa.subatomic.infrastructure.member.view.jpa.TeamMemberEntity;
 import za.co.absa.subatomic.infrastructure.member.view.jpa.TeamMemberPersistenceHandler;
 import za.co.absa.subatomic.infrastructure.project.view.jpa.ProjectEntity;
@@ -49,10 +47,6 @@ public class TeamService {
     private MembershipRequestRepository membershipRequestRepository;
 
     private TeamAssertions teamAssertions = new TeamAssertions();
-
-    private RestTemplate restTemplate;
-
-    private AtomistConfigurationProperties atomistConfigurationProperties;
 
     public TeamService(TeamRepository teamRepository,
             MembershipRequestRepository membershipRequestRepository,
@@ -149,22 +143,33 @@ public class TeamService {
                 actionedByEntity);
     }
 
-    public TeamEntity addSlackIdentity(String teamId, String teamChannel) {
+    public TeamEntity addSlackIdentity(String teamId, String actionedByMemberId, String teamChannel) {
+        // actionedByMemberId (is the now the variable name standard) = createBy = requestedById
+        TeamEntity teamEntity = this.persistenceHandler.addSlackIdentity(teamId,
+                teamChannel);
 
-        AtomistTeam atomistTeam = new AtomistTeam (newTeamEntity.getTeamId(),
-                newTeamEntity.getName(),
-                newTeamEntity.getOpenShiftCloud(),
+        TeamMemberEntity member = this.teamMemberPersistenceHandler
+                .findByTeamMemberId(actionedByMemberId);
+
+        TeamSlackIdentity teamSlackIdentity = new TeamSlackIdentity(
+                teamEntity.getSlackDetails().getTeamChannel());
+
+        TeamMemberSlackIdentity teamMemberSlackIdentity = new TeamMemberSlackIdentity(
+                member.getSlackDetails().getScreenName(),
+                member.getSlackDetails().getUserId());
+
+        AtomistTeam atomistTeam = new AtomistTeam (teamEntity.getTeamId(),
+                teamEntity.getName(),
+                teamEntity.getOpenShiftCloud(),
                 teamSlackIdentity);
 
-        AtomistMemberBase atomistMemberBase = new AtomistMemberBase(
-                teamMemberEntity.getFirstName(),
-                teamMemberEntity.getDomainUsername(),
+         AtomistMemberBase atomistMemberBase = new AtomistMemberBase(
+                member.getFirstName(),
+                member.getDomainUsername(),
                 teamMemberSlackIdentity);
 
-        teamSlackChannelCreated(atomistTeam,atomistMemberBase);
-
-        return this.persistenceHandler.addSlackIdentity(teamId,
-                teamChannel);
+        this.automationHandler.teamSlackChannelCreated(atomistTeam,atomistMemberBase);
+        return teamEntity;
     }
 
     public void newDevOpsEnvironment(String teamId, String requestedById) {
@@ -256,26 +261,6 @@ public class TeamService {
                 requestedByMemberEntity, previousCloud);
     }
 
-    public void teamSlackChannelCreated(AtomistTeam atomistTeam, AtomistMemberBase atomistMemberBase) {
-        // Raise event for teamSlackChannelCreated -- todo: move to team service as a function and call from here
-
-        TeamService.TeamSlackChannelCreated ingestableObject = new TeamService.TeamSlackChannelCreated(
-                atomistTeam, atomistMemberBase);
-
-        ResponseEntity<String> responseTeamSlackChannelCreated = restTemplate.postForEntity(
-                atomistConfigurationProperties
-                        .getMemberRemovedFromTeamEventUrl(),
-                ingestableObject,
-                String.class);
-
-        if (responseTeamSlackChannelCreated.getStatusCode().is2xxSuccessful()) {
-            log.info("Atomist has ingested TeamSlackChannelCreatedEvent successfully: -> {}",
-                    responseTeamSlackChannelCreated.getHeaders(),
-                    responseTeamSlackChannelCreated.getBody());
-        }
-        // End of event TeamSlackChannelCreated
-    }
-
     @Transactional(readOnly = true)
     public Set<TeamEntity> findTeamsAssociatedToProject(String projectId) {
         ProjectEntity projectEntity = projectRepository
@@ -284,13 +269,6 @@ public class TeamService {
         associatedTeams.add(projectEntity.getOwningTeam());
         associatedTeams.addAll(projectEntity.getTeams());
         return associatedTeams;
-    }
-
-    @Value
-    private class TeamSlackChannelCreated {
-        private AtomistTeam team;
-
-        private AtomistMemberBase createdBy;
     }
 
     public TeamPersistenceHandler getPersistenceHandler() {
